% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcGridGeometries.R
\name{calcGridGeometries}
\alias{calcGridGeometries}
\title{Calculates grid lat, lon and area}
\usage{
calcGridGeometries(grid)
}
\arguments{
\item{grid}{Raster grid including projection information}
}
\value{
nothing, just save the generated rasters as GeoTiff
}
\description{
Generate three raster layers containing pixels longitude, latitude and
area information and save it to disk.
}

