% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcOccurrenceMetrics.R
\name{calcOccurrenceMetrics}
\alias{calcOccurrenceMetrics}
\title{Calculate metrics for species occurence}
\usage{
calcOccurrenceMetrics(species_polygon)
}
\arguments{
\item{species_polygon}{Spatial polygon containing information about a single
species}
}
\value{
List object containing subareals and centroids as spatial objects and
a variety of metrics.
}
\description{
Calculate metrics for species occurence
}

