% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPointDist.R
\name{calcPointDist}
\alias{calcPointDist}
\title{Calculates distance between points on a sphere}
\usage{
calcPointDist(points)
}
\arguments{
\item{points}{Spatial points object}
}
\value{
Table containing the distances between each points in meter.
}
\description{
Calculates distance between points using distVincentyEllipsoid method
}

