% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPointSubAreals.R
\name{calcPointSubAreals}
\alias{calcPointSubAreals}
\title{Calculates sub areals}
\usage{
calcPointSubAreals(points)
}
\arguments{
\item{points}{a SpatialPointsDataFrame in a metric coordinate system}
}
\value{
Spatial polygon containing the sub areals
}
\description{
Calculates buffers around individual observations to create sub areals for
the individual
}
\details{
The areals are based on multiple buffer computations to ensure
that the required vector length is still allocatable.
}

