% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPolygonArea.R
\name{calcPolygonArea}
\alias{calcPolygonArea}
\title{Calculate surface of polygons on a cartographic grid}
\usage{
calcPolygonArea(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{a Spatial polygon* in a metric coordinate system}
}
\value{
Data frame containing the overall area, the area without holes,
the area of the holes (each information absolute and relative)
}
\description{
Calculates surface of polygons based on non-spherical coordinates.
}

