% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPolygonAreaByID.R
\name{calcPolygonAreaByID}
\alias{calcPolygonAreaByID}
\title{Calculate individual surface of multiple polygons on a cartographic grid}
\usage{
calcPolygonAreaByID(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{a Spatial polygon* in a metric coordinate system}
}
\value{
Data frame containing the area of each sub-polygon and information if
this area is for a hole or not.
}
\description{
Calculates surface of polygons based on non-spherical coordinates.
}

