% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPolygonAreaSphere.R
\name{calcPolygonAreaSphere}
\alias{calcPolygonAreaSphere}
\title{Calculate surface of triangular polygons on the earth sphere}
\usage{
calcPolygonAreaSphere(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{a Spatial polygon* in a spherical coordinate system}
}
\value{
Data frame containing the area of all sub-polygons.
}
\description{
Calculates surface of polygons based on spherical coordinates.
}

