% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPolygonAreaSphereByID.R
\name{calcPolygonAreaSphereByID}
\alias{calcPolygonAreaSphereByID}
\title{Calculate surface of multiple triangular polygons on the earth sphere}
\usage{
calcPolygonAreaSphereByID(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{a Spatial polygon* in a spherical coordinate system}
}
\value{
Data frame containing the area of each sub-polygons.
}
\description{
Calculates surface of polygons based on spherical coordinates.
}

