% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPolygonCentroids.R
\name{calcPolygonCentroids}
\alias{calcPolygonCentroids}
\title{Calculate centroids of polygons}
\usage{
calcPolygonCentroids(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{Spatial polygon}
}
\value{
Spatial.Polygon* object
}
\description{
Calculate centroids of polygons.
}

