% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dividePolygon.R
\name{dividePolygon}
\alias{dividePolygon}
\title{Divide spatial polygon into triangles}
\usage{
dividePolygon(spatial_polygon)
}
\arguments{
\item{spatial_polygon}{SpatialPolygon* object}
}
\value{
SpatialPolygon* object
}
\description{
Divide spatial polygon into triangles.
}

