% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geocodeByCountry.R
\name{geocodeByCountry}
\alias{geocodeByCountry}
\title{Geocode dataset by country code}
\usage{
geocodeByCountry(data, col = "countryCode", country)
}
\arguments{
\item{data}{Dataset to be updated}

\item{col}{Column containing the country code information}

\item{country}{Dataset containing country codes and centroid coordinates}
}
\value{
data.table object
}
\description{
Assign centroid coordinates by country code
}

