% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readLTF.R
\name{readLTF}
\alias{readLTF}
\title{Read large structure text file}
\usage{
readLTF(infile, sep = ",", header = TRUE, rlvt_cols = NULL)
}
\arguments{
\item{sep}{Separator of the text file columns}

\item{header}{Does the first line contains a header (TRUE/FALSE)}

\item{rlvt_cols}{Include only given columns in the output data.table.}

\item{filename}{Name(s) of the data file}
}
\value{
data.table object
}
\description{
Read large structure text file from chunks or as a whole.
}

