% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speciesdist.R
\name{speciesdist}
\alias{speciesdist}
\title{Species distribution}
\usage{
speciesdist()
}
\description{
Methods for analysing species distributions on a global scale


}
\details{
The package provides usefull functions for species analysis on a
global scale.
}

