% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splitLTF.R
\name{splitLTF}
\alias{splitLTF}
\title{Split large structure text file}
\usage{
splitLTF(infile, sep, header = TRUE, outpath)
}
\arguments{
\item{sep}{Separator of the text file columns}

\item{outpath}{Name of the output directory where the individual chunks are
stored.}

\item{filename}{Name of the data file}

\item{sep}{Does the first line contains a header (TRUE/FALSE)}
}
\value{
List object containing some meta information on the processing.
}
\description{
Divide large structure text file into chunks of 100,000 lines and check if
each line has the same amount of lines.
}

